+++ src/Crowdin.Api/Glossaries/Glossary.cs
@@ -15,7 +15,7 @@ namespace Crowdin.Api.Glossaries
         public string Name { get; set; }
         
         [JsonProperty("groupId")]
-        public int? GroupId { get; set; }
+        public long? GroupId { get; set; }
         
         [JsonProperty("userId")]
         public long UserId { get; set; }
@@ -30,10 +30,10 @@ namespace Crowdin.Api.Glossaries
         public string[] LanguageIds { get; set; }
         
         [JsonProperty("defaultProjectIds")]
-        public int[] DefaultProjectIds { get; set; }
+        public long[] DefaultProjectIds { get; set; }
         
         [JsonProperty("projectIds")]
-        public int[] ProjectIds { get; set; }
+        public long[] ProjectIds { get; set; }
         
         [JsonProperty("createdAt")]
         public DateTimeOffset CreatedAt { get; set; }
