+++ src/Crowdin.Api/SourceFiles/FileResource.cs
@@ -17,10 +17,10 @@ namespace Crowdin.Api.SourceFiles
         public long ProjectId { get; set; }
         
         [JsonProperty("branchId")]
-        public int? BranchId { get; set; }
+        public long? BranchId { get; set; }
         
         [JsonProperty("directoryId")]
-        public int? DirectoryId { get; set; }
+        public long? DirectoryId { get; set; }
         
         [JsonProperty("name")]
 #pragma warning disable CS8618
